+++ src/main/groovy/com/rapidminer/gradle/RapidMinerCodeQualityPlugin.groovy
@@ -18,14 +18,21 @@ package com.rapidminer.gradle
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.Task
-import org.gradle.api.plugins.quality.PmdPlugin
+import org.gradle.api.plugins.GroovyPlugin
+import org.gradle.api.plugins.JavaPlugin
+import org.gradle.plugins.ide.eclipse.EclipsePlugin
 import org.gradle.testing.jacoco.plugins.JacocoPlugin
 
 import com.rapidminer.gradle.checkstyle.InitCheckstyleConfigFiles
 import com.rapidminer.gradle.codenarc.InitCodenarcConfigFiles
+import com.rapidminer.gradle.eclipse.checkstyle.CheckstyleEclipse
+import com.rapidminer.gradle.eclipse.findbugs.FindbugsEclipse
+import com.rapidminer.gradle.eclipse.pmd.PMDEclipse
+
 
 
 /**
+ * The Code Quality plugin class which contains all logic for applying code quality plugins.
  *
  * @author Nils Woehler
  *
@@ -36,18 +43,23 @@ class RapidMinerCodeQualityPlugin implements Plugin<Project> {
 	private static final String ALL_JAVA = '**/*.java'
 	private static final String ALL_GROOVY = '**/*.groovy'
 
+	private static final String HEADER_CHECK = 'headerCheck'
+	private static final String JDEPEND = 'jdepend'
+	private static final String JACOCO = 'jacoco'
+
+	private static final String PMD = 'pmd'
+	private static final String ECLIPSE_PMD_CONFIG_TASK = 'eclipsePMD'
+
 	private static final String CHECKSTYLE = 'checkstyle'
+	private static final String ECLIPSE_CHECKSTYLE_CONFIG_TASK = 'eclipseCheckstyle'
 	private static final String INIT_CHECKSTYLE_CONFIG_TASK = 'checkstyleInitDefaultConfig'
 	private static final String CHECKSTYLE_GEN_REPORT = 'HTMLReport'
 
 	private static final String CODENARC = 'codenarc'
 	private static final String INIT_CODENARC_CONFIG_TASK = 'codenarcInitDefaultConfig'
 
-	private static final String HEADER_CHECK = 'headerCheck'
-	private static final String JDEPEND = 'jdepend'
 	private static final String FINDBUGS = 'findbugs'
-	private static final String JACOCO = 'jacoco'
-	private static final String PMD = 'pmd'
+	private static final String ECLIPSE_FINDBUGS_CONFIG_TASK = 'eclipseFindBugs'
 
 	@Override
 	void apply(Project project) {
@@ -56,9 +68,9 @@ class RapidMinerCodeQualityPlugin implements Plugin<Project> {
 		CodeQualityConfiguration qualityExt = project.extensions.create("codeQuality", CodeQualityConfiguration)
 
 		// Adds JaCoCo check tasks
-		// Cannot be done in afterEvaluate as this breaks the plugin tasks
-		if(applyPlugin(project, JACOCO, qualityExt.jacoco) && 
-			(project.plugins.withType(org.gradle.api.plugins.GroovyPlugin) || project.plugins.withType(org.gradle.api.plugins.JavaPlugin))){
+		// Cannot be done in afterEvaluate as this breaks the JaCoCo plugin tasks
+		if(applyPlugin(project, JACOCO, qualityExt.jacoco) &&
+		(project.plugins.withType(org.gradle.api.plugins.GroovyPlugin) || project.plugins.withType(org.gradle.api.plugins.JavaPlugin))){
 			configureJaCoCo(project, qualityExt)
 		}
 
@@ -70,7 +82,7 @@ class RapidMinerCodeQualityPlugin implements Plugin<Project> {
 				configureHeaderCheck(project, qualityExt, configurationDir)
 			}
 
-			if(project.plugins.withType(org.gradle.api.plugins.GroovyPlugin)) {
+			if(project.plugins.withType(GroovyPlugin)) {
 				project.logger.info("${project.name} is a Groovy project. Only checking for Groovy code quality plugins.")
 
 				if(applyPlugin(project, CODENARC, qualityExt.codenarc)) {
@@ -78,7 +90,7 @@ class RapidMinerCodeQualityPlugin implements Plugin<Project> {
 					configureCodeNarc(project, qualityExt, configurationDir)
 				}
 
-			} else if(project.plugins.withType(org.gradle.api.plugins.JavaPlugin)) {
+			} else if(project.plugins.withType(JavaPlugin)) {
 				project.logger.info("${project.name} is a Java project. Only checking for Java code quality plugins.")
 
 				if(applyPlugin(project, CHECKSTYLE,  qualityExt.checkstyle)) {
@@ -126,12 +138,38 @@ class RapidMinerCodeQualityPlugin implements Plugin<Project> {
 	private void configurePMD(Project project, CodeQualityConfiguration codeExt) {
 		project.configure(project){
 			apply plugin: 'pmd'
-		
+
+			File ruleSet = project.rootProject.file(codeExt.configDir + '/pmd/ruleset.xml')
+
 			pmd {
 				ignoreFailures = { codeExt.pmdIgnoreErrors }
 				sourceSets = [sourceSets.main]
+				ruleSetFiles = files(ruleSet)
+				
+				// disable predfined basic ruleset
+				ruleSets = []
 			}
-			
+
+			// Create task which allows to configure PMD Eclipse plugin
+			Task initPMDEclipseTask = tasks.create(name: ECLIPSE_PMD_CONFIG_TASK, type: PMDEclipse)
+			initPMDEclipseTask.group = TASK_GROUP
+			initPMDEclipseTask.description = "Creates PMD Eclipse plugin config files for the current project."
+
+			project.tasks.findAll { return it.name.startsWith('pmd') }.each { it.dependsOn initPMDEclipseTask }
+
+			initPMDEclipseTask.configure { rulesetFile = ruleSet }
+
+			// Ensure findbugs config files are copied if project applies Eclipse plugin
+			project.plugins.withType(EclipsePlugin) {
+				project.tasks.eclipse.dependsOn initPMDEclipseTask
+
+				// Also configure PMD nature and buildCommand
+				eclipse.project {
+					natures 'net.sourceforge.pmd.eclipse.plugin.pmdNature'
+					buildCommand 'net.sourceforge.pmd.eclipse.plugin.pmdBuilder'
+				}
+			}
+
 		}
 	}
 
@@ -158,8 +196,36 @@ class RapidMinerCodeQualityPlugin implements Plugin<Project> {
 				ignoreFailures = { codeExt.findbugsIgnoreErrors }
 				reportsDir = file("${project.buildDir}/reports/findbugs")
 
+				if(file(codeExt.findbugsExcludeFilter).exists()){
+					excludeFilter = file(codeExt.findbugsExcludeFilter)
+				}
+
 				effort = 'max'
-				reportLevel = 'high'
+				reportLevel = 'medium'
+			}
+
+
+			// check if provided configuration exists
+			if(project.configurations.find { it.name == 'provided' }) {
+				dependencies { provided 'com.google.code.findbugs:annotations:3.0.0' }
+			} else {
+				project.logger.info('Configuration \'provided\' does not exist. Skip adding of FindBugs annotation library dependency.')
+			}
+
+			// Create task which allows to configure FindBugs Eclipse plugin
+			Task initFindbugsEclipseTask = tasks.create(name: ECLIPSE_FINDBUGS_CONFIG_TASK, type: FindbugsEclipse)
+			initFindbugsEclipseTask.group = TASK_GROUP
+			initFindbugsEclipseTask.description = "Creates FindBugs Eclipse plugin config files for the current project."
+
+			// Ensure findbugs config files are copied if project applies Eclipse plugin
+			project.plugins.withType(EclipsePlugin) {
+				project.tasks.eclipse.dependsOn initFindbugsEclipseTask
+
+				// Also configure FindBugs nature and buildCommand
+				eclipse.project {
+					natures 'edu.umd.cs.findbugs.plugin.eclipse.findbugsNature'
+					buildCommand 'edu.umd.cs.findbugs.plugin.eclipse.findbugsBuilder'
+				}
 			}
 		}
 	}
@@ -187,9 +253,9 @@ class RapidMinerCodeQualityPlugin implements Plugin<Project> {
 			apply plugin: CODENARC
 
 			// ensure codenarc config file is in place
-			tasks.create(name: INIT_CODENARC_CONFIG_TASK, type: InitCodenarcConfigFiles)
-			codenarcInitDefaultConfig.group = TASK_GROUP
-			codenarcInitDefaultConfig.description = "Copies the default codenarc.conf files to " +
+			Task initTask = tasks.create(name: INIT_CODENARC_CONFIG_TASK, type: InitCodenarcConfigFiles)
+			initTask.group = TASK_GROUP
+			initTask.description = "Copies the default codenarc.conf files to " +
 					"the configured configuration directory."
 
 			def codenarcConfigDir = new File(configurationDir.absolutePath, CODENARC)
@@ -218,8 +284,8 @@ class RapidMinerCodeQualityPlugin implements Plugin<Project> {
 		project.configure(project) {
 			apply plugin: CHECKSTYLE
 
-			def checkstyleConfigDir = new File(configurationDir.absolutePath, CHECKSTYLE)
-			def checkstyleConfigFile = new File(checkstyleConfigDir.absolutePath, ext.checkstyleConfigFileName)
+			File checkstyleConfigDir = new File(configurationDir.absolutePath, CHECKSTYLE)
+			File checkstyleConfigFile = new File(checkstyleConfigDir.absolutePath, ext.checkstyleConfigFileName)
 
 			checkstyleMain {
 				ignoreFailures = ext.checkstyleIgnoreErrors
@@ -254,23 +320,50 @@ class RapidMinerCodeQualityPlugin implements Plugin<Project> {
 			}
 
 			// ensure checkstyle config files are in place
-			tasks.create(name: INIT_CHECKSTYLE_CONFIG_TASK, type: InitCheckstyleConfigFiles)
-			checkstyleInitDefaultConfig.group = TASK_GROUP
-			checkstyleInitDefaultConfig.description = "Copies the default checkstyle.xml files to " +
+			Task generateCheckstyleConfigTask = tasks.create(name: INIT_CHECKSTYLE_CONFIG_TASK, type: InitCheckstyleConfigFiles)
+			generateCheckstyleConfigTask.group = TASK_GROUP
+			generateCheckstyleConfigTask.description = "Copies the default checkstyle.xml files to " +
 					"the configured configuration directory."
 
-			// Configure checkstyle tasks
+			// Configure checkstyle init config tasks
 			if(ext.checkstyleUseDefaultConfig) {
 				project.tasks.each { t ->
 					if(t.name != INIT_CHECKSTYLE_CONFIG_TASK && t.name.startsWith(CHECKSTYLE)) {
-						t.dependsOn checkstyleInitDefaultConfig
+						t.dependsOn generateCheckstyleConfigTask
 					}
 				}
 				checkstyleInitDefaultConfig {
 					configDir = checkstyleConfigDir
 					checkstyleFileName = ext.checkstyleConfigFileName
 				}
+
+				// if default config should be used, also configure Eclipse plugin config file task
+				Task eclipseCheckstyle = tasks.create(name: ECLIPSE_CHECKSTYLE_CONFIG_TASK, type: CheckstyleEclipse)
+				eclipseCheckstyle.group = TASK_GROUP
+				eclipseCheckstyle.description = "Creates Checkstyle Eclipse plugin config files for the current project."
+
+				eclipseCheckstyle.configure {
+					configDir = ext.configDir
+					checkstyleFileName = ext.checkstyleConfigFileName
+				}
+
+				// Ensure findbugs config files are copied if project applies Eclipse plugin
+				project.plugins.withType(EclipsePlugin) {
+					project.tasks.eclipse.dependsOn eclipseCheckstyle
+
+					// ensure that checkstyle.xml is present
+					eclipseCheckstyle.dependsOn generateCheckstyleConfigTask
+
+					// Also configure checkstyle nature and buildCommand
+					eclipse.project {
+						natures 'net.sf.eclipsecs.core.CheckstyleNature'
+						buildCommand 'net.sf.eclipsecs.core.CheckstyleBuilder'
+					}
+				}
 			}
+
+
+
 		}
 	}